requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
Django==4.0.4
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.11.1
flake8==4.0.1
flake8-html==0.4.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
sentry-sdk==1.5.12
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
